requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres

#apache-airflow-providers-google==21.0.0
astro-sdk-python[google]>=1.8.0
#soda-core-bigquery==3.5.6

astronomer-cosmos[dbt-bigquery]
openlineage-integration-common==1.26.0
protobuf
openpyxl
